requirements.txt
django==4.2.3
djangorestframework==3.14.0
pytest-django==4.5.2
drf-yasg==1.21.6
djangorestframework-simplejwt==5.2.2
django-rest-passwordreset==1.3.0
autoflake==2.2.0
black==23.3.0
isort==5.12.0
flake8==6.0.0
flake8-bugbear==23.6.5
flake8-comprehensions==3.13.0
flake8-eradicate==1.5.0
flake8-pie==0.16.0
pyflakes==3.0.1